Core AMQP cleanup for v2
