+++ sdk/core/core-amqp/test/config.spec.ts
@@ -1,10 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ConnectionConfig, EventHubConnectionConfig, IotHubConnectionConfig } from "../src";
+import { ConnectionConfig } from "../src";
 import * as chai from "chai";
 import { isSharedAccessSignature } from "../src/connectionConfig/connectionConfig";
-import { SharedAccessSignatureCredential } from "../src/auth/sas";
 const should = chai.should();
 
 describe("ConnectionConfig", function() {
@@ -197,241 +196,6 @@ describe("ConnectionConfig", function() {
     });
   });
 
-  describe("EventHub", function() {
-    it("should fail if connection config does not contain path and the connectionstring also does not contain EntityPath", function(done) {
-      const connectionString =
-        "Endpoint=sb://hostname.servicebus.windows.net/;SharedAccessKeyName=sakName;SharedAccessKey=sak";
-      try {
-        EventHubConnectionConfig.create(connectionString);
-        done(new Error("Should not have reached here."));
-      } catch (err) {
-        err.message.should.match(/Either provide "path" or the "connectionString".*/gi);
-      }
-      done();
-    });
-
-    it("should correctly populate config properties from an EventHubs connection string and the helper methods should work as expected", function(done) {
-      const config = EventHubConnectionConfig.create(
-        "Endpoint=sb://hostname.servicebus.windows.net/;SharedAccessKeyName=sakName;SharedAccessKey=sak;EntityPath=ep"
-      );
-      config.should.have.property("host").that.equals("hostname.servicebus.windows.net");
-      config.should.have.property("sharedAccessKeyName").that.equals("sakName");
-      config.should.have.property("sharedAccessKey").that.equals("sak");
-      config.should.have.property("entityPath").that.equals("ep");
-
-      config.getManagementAddress().should.equal("ep/$management");
-      config.getSenderAddress().should.equal("ep");
-      config.getSenderAddress("0").should.equal("ep/Partitions/0");
-      config.getSenderAddress(0).should.equal("ep/Partitions/0");
-      config.getReceiverAddress("0").should.equal("ep/ConsumerGroups/$default/Partitions/0");
-      config.getReceiverAddress(0).should.equal("ep/ConsumerGroups/$default/Partitions/0");
-      config.getReceiverAddress("0", "cg").should.equal("ep/ConsumerGroups/cg/Partitions/0");
-      config.getReceiverAddress(0, "cg").should.equal("ep/ConsumerGroups/cg/Partitions/0");
-
-      config
-        .getManagementAudience()
-        .should.equal("sb://hostname.servicebus.windows.net/ep/$management");
-      config.getSenderAudience().should.equal("sb://hostname.servicebus.windows.net/ep");
-      config
-        .getSenderAudience("0")
-        .should.equal("sb://hostname.servicebus.windows.net/ep/Partitions/0");
-      config
-        .getSenderAudience(0)
-        .should.equal("sb://hostname.servicebus.windows.net/ep/Partitions/0");
-      config
-        .getReceiverAudience("0")
-        .should.equal(
-          "sb://hostname.servicebus.windows.net/ep/ConsumerGroups/$default/Partitions/0"
-        );
-      config
-        .getReceiverAudience(0)
-        .should.equal(
-          "sb://hostname.servicebus.windows.net/ep/ConsumerGroups/$default/Partitions/0"
-        );
-      config
-        .getReceiverAudience("0", "cg")
-        .should.equal("sb://hostname.servicebus.windows.net/ep/ConsumerGroups/cg/Partitions/0");
-      config
-        .getReceiverAudience(0, "cg")
-        .should.equal("sb://hostname.servicebus.windows.net/ep/ConsumerGroups/cg/Partitions/0");
-      done();
-    });
-
-    it("requires that Endpoint be present in the connection string", (done) => {
-      const connectionString = `Endpoint=sb://a`;
-
-      should.throw(() => {
-        EventHubConnectionConfig.create(connectionString);
-      }, /must contain EntityPath/);
-
-      done();
-    });
-  });
-
-  describe("IotHub", function() {
-    const iotString =
-      "HostName=someiot.azure-devices.net;SharedAccessKeyName=sakName;SharedAccessKey=sak;DeviceId=device-1234";
-    it("should correctly create an iothub connection config from an iothub connectionstring", function(done) {
-      const config = IotHubConnectionConfig.create(iotString);
-      config.should.have.property("hostName").that.equals("someiot.azure-devices.net");
-      config.should.have.property("host").that.equals("someiot");
-      config.should.have.property("sharedAccessKeyName").that.equals("sakName");
-      config.should.have.property("sharedAccessKey").that.equals("sak");
-      config.should.have.property("entityPath").that.equals("messages/events");
-      done();
-    });
-
-    it("populates path from the path argument if provided", function(done) {
-      const config = IotHubConnectionConfig.create(iotString, "abc");
-      config.should.have.property("entityPath").that.equals("abc");
-      done();
-    });
-
-    it("converts an IotHubConnectionConfig to an EventHubConnectionConfig", function(done) {
-      const config = IotHubConnectionConfig.create(iotString);
-      const ehConfig = IotHubConnectionConfig.convertToEventHubConnectionConfig(config);
-      ehConfig.should.have.property("endpoint").that.equals("sb://someiot.azure-devices.net/");
-      ehConfig.should.have.property("host").that.equals("someiot.azure-devices.net");
-      ehConfig.should.have.property("sharedAccessKeyName").that.equals("sakName");
-      ehConfig.should.have.property("sharedAccessKey").that.equals("sak");
-      ehConfig.should.have.property("entityPath").that.equals("messages/events");
-
-      ehConfig.getManagementAddress().should.equal("messages/events/$management");
-      ehConfig.getSenderAddress().should.equal("messages/events");
-      ehConfig.getSenderAddress("0").should.equal("messages/events/Partitions/0");
-      ehConfig.getSenderAddress(0).should.equal("messages/events/Partitions/0");
-      ehConfig
-        .getReceiverAddress("0")
-        .should.equal("messages/events/ConsumerGroups/$default/Partitions/0");
-      ehConfig
-        .getReceiverAddress(0)
-        .should.equal("messages/events/ConsumerGroups/$default/Partitions/0");
-      ehConfig
-        .getReceiverAddress("0", "cg")
-        .should.equal("messages/events/ConsumerGroups/cg/Partitions/0");
-      ehConfig
-        .getReceiverAddress(0, "cg")
-        .should.equal("messages/events/ConsumerGroups/cg/Partitions/0");
-
-      ehConfig
-        .getManagementAudience()
-        .should.equal("sb://someiot.azure-devices.net/messages/events/$management");
-      ehConfig.getSenderAudience().should.equal("sb://someiot.azure-devices.net/messages/events");
-      ehConfig
-        .getSenderAudience("0")
-        .should.equal("sb://someiot.azure-devices.net/messages/events/Partitions/0");
-      ehConfig
-        .getSenderAudience(0)
-        .should.equal("sb://someiot.azure-devices.net/messages/events/Partitions/0");
-      ehConfig
-        .getReceiverAudience("0")
-        .should.equal(
-          "sb://someiot.azure-devices.net/messages/events/ConsumerGroups/$default/Partitions/0"
-        );
-      ehConfig
-        .getReceiverAudience(0)
-        .should.equal(
-          "sb://someiot.azure-devices.net/messages/events/ConsumerGroups/$default/Partitions/0"
-        );
-      ehConfig
-        .getReceiverAudience("0", "cg")
-        .should.equal(
-          "sb://someiot.azure-devices.net/messages/events/ConsumerGroups/cg/Partitions/0"
-        );
-      ehConfig
-        .getReceiverAudience(0, "cg")
-        .should.equal(
-          "sb://someiot.azure-devices.net/messages/events/ConsumerGroups/cg/Partitions/0"
-        );
-      done();
-    });
-
-    describe("Throws error if required connection config properties are not present", function() {
-      const connectionString = `
-        Endpoint=sb://someiot.azure-devices.net/;
-        SharedAccessKeyName=sakName;
-        SharedAccessKey=sakName;
-        EntityPath=messages/events;
-      `;
-
-      it("requires that connection config be present", (done) => {
-        should.throw(() => {
-          IotHubConnectionConfig.validate(undefined as any);
-        }, /Missing configuration/);
-
-        done();
-      });
-
-      it("requires that hostName be present in the connection config", (done) => {
-        const config: IotHubConnectionConfig = {
-          connectionString: connectionString,
-          hostName: "",
-          host: "someiot",
-          sharedAccessKeyName: "sakName",
-          sharedAccessKey: "sak",
-          entityPath: "messages/events"
-        };
-
-        should.throw(() => {
-          IotHubConnectionConfig.validate(config);
-        }, /Missing 'hostName'/);
-
-        done();
-      });
-
-      it("requires that host be present in the connection config", (done) => {
-        const config: IotHubConnectionConfig = {
-          connectionString: connectionString,
-          hostName: "someiot.azure-devices.net",
-          host: "someiot",
-          sharedAccessKeyName: "sakName",
-          sharedAccessKey: "sak",
-          entityPath: ""
-        };
-
-        should.throw(() => {
-          IotHubConnectionConfig.validate(config);
-        }, /Missing 'entityPath'/);
-
-        done();
-      });
-
-      it("requires that sharedAccessKeyName be present in the connection config", (done) => {
-        const config: IotHubConnectionConfig = {
-          connectionString: connectionString,
-          hostName: "someiot.azure-devices.net",
-          host: "someiot",
-          sharedAccessKeyName: "",
-          sharedAccessKey: "sak",
-          entityPath: "messages/events"
-        };
-
-        should.throw(() => {
-          IotHubConnectionConfig.validate(config);
-        }, /Missing 'sharedAccessKeyName'/);
-
-        done();
-      });
-
-      it("requires that sharedAccessKey be present in the connection config", (done) => {
-        const config: IotHubConnectionConfig = {
-          connectionString: connectionString,
-          hostName: "someiot.azure-devices.net",
-          host: "someiot",
-          sharedAccessKeyName: "sakName",
-          sharedAccessKey: "",
-          entityPath: "messages/events"
-        };
-
-        should.throw(() => {
-          IotHubConnectionConfig.validate(config);
-        }, /Missing 'sharedAccessKey'/);
-
-        done();
-      });
-    });
-  });
-
   describe("SharedAccessSignature", () => {
     [
       "Endpoint=hello;SharedAccessSignature=SharedAccessSignature sr=<resource>&sig=someb64=&se=<expiry>&skn=<keyname>",
@@ -460,29 +224,6 @@ describe("ConnectionConfig", function() {
         sharedAccessKeyName: "",
         connectionString: "Endpoint=hello;SharedAccessSignature=SharedAccessSignature hellosig"
       });
-    });
-
-    it("SharedAccessSignatureCredential", () => {
-      const sasCred = new SharedAccessSignatureCredential("SharedAccessSignature se=<blah>");
-      const accessToken = sasCred.getToken("audience isn't used");
-
-      should.equal(
-        accessToken.token,
-        "SharedAccessSignature se=<blah>",
-        "SAS URI we were constructed with should just be returned verbatim without interpretation (and the audience is ignored)"
-      );
-
-      should.equal(
-        accessToken.expiresOnTimestamp,
-        0,
-        "SAS URI always returns 0 for expiry (ignoring what's in the SAS token)"
-      );
-
-      // these just exist because we're a SharedKeyCredential but we don't currently
-      // parse any attributes out (they're available but we've carved out a spot so
-      // they're not needed.)
-      should.equal(sasCred.key, "");
-      should.equal(sasCred.keyName, "");
     });
   });
 });
