+++ sdk/core/core-amqp/test/context.spec.ts
@@ -3,13 +3,7 @@
 
 import * as chai from "chai";
 const should = chai.should();
-import {
-  CbsClient,
-  ConnectionConfig,
-  ConnectionContextBase,
-  DefaultDataTransformer,
-  SharedKeyCredential
-} from "../src";
+import { CbsClient, ConnectionConfig, ConnectionContextBase, DefaultDataTransformer } from "../src";
 import { Connection } from "rhea-promise";
 import { isNode } from "../src/util/utils";
 
@@ -32,10 +26,8 @@ describe("ConnectionContextBase", function() {
     should.exist(context.connectionId);
     should.exist(context.connectionLock);
     should.exist(context.negotiateClaimLock);
-    should.exist(context.tokenCredential);
     should.exist(context.dataTransformer);
     context.wasConnectionCloseCalled.should.equal(false);
-    context.tokenCredential.should.instanceOf(SharedKeyCredential);
     context.connection.should.instanceOf(Connection);
     context.connection.options.properties!.product.should.equal("MSJSClient");
     context.connection.options.properties!["user-agent"].should.equal("/js-amqp-client");
@@ -141,7 +133,6 @@ describe("ConnectionContextBase", function() {
       should.exist(context.connectionId);
       should.exist(context.connectionLock);
       should.exist(context.negotiateClaimLock);
-      should.exist(context.tokenCredential);
       should.exist(context.dataTransformer);
       context.wasConnectionCloseCalled.should.equal(false);
       context.cbsSession.should.instanceOf(CbsClient);
