+++ sdk/eventhub/event-hubs/package.json
@@ -91,9 +91,12 @@
     "@azure/core-amqp": "2.0.0-beta.1",
     "@azure/core-asynciterator-polyfill": "^1.0.0",
     "@azure/core-tracing": "1.0.0-preview.9",
+    "@azure/core-auth": "^1.1.3",
     "@azure/logger": "^1.0.0",
     "@opentelemetry/api": "^0.10.2",
     "buffer": "^5.2.1",
+    "is-buffer": "^2.0.3",
+    "jssha": "^3.1.0",
     "process": "^0.11.10",
     "rhea-promise": "^1.0.0",
     "tslib": "^2.0.0",
