+++ sdk/eventhub/event-hubs/test/diagnostics/messageSpan.spec.ts
@@ -5,7 +5,7 @@ import chai from "chai";
 import { createMessageSpan } from "../../src/diagnostics/messageSpan";
 import { TraceFlags, SpanContext } from "@opentelemetry/api";
 import { TestTracer, setTracer, getTracer } from "@azure/core-tracing";
-import { EventHubConnectionConfig } from "@azure/core-amqp";
+import { EventHubConnectionConfig } from "../../src/eventhubConnectionConfig";
 
 const should = chai.should();
 const assert = chai.assert;
