+++ sdk/servicebus/service-bus/package.json
@@ -102,12 +102,14 @@
     "@azure/core-http": "^1.2.0",
     "@azure/core-tracing": "1.0.0-preview.9",
     "@azure/core-paging": "^1.1.1",
+    "@azure/core-auth": "^1.1.3",
     "@azure/logger": "^1.0.0",
     "@opentelemetry/api": "^0.10.2",
     "@types/is-buffer": "^2.0.0",
     "@types/long": "^4.0.0",
     "buffer": "^5.2.1",
     "is-buffer": "^2.0.3",
+    "jssha": "^3.1.0",
     "long": "^4.0.0",
     "process": "^0.11.10",
     "tslib": "^2.0.0",
