+++ sdk/servicebus/service-bus/src/constructorHelpers.ts
@@ -4,12 +4,12 @@
 import {
   ConnectionConfig,
   RetryOptions,
-  SharedKeyCredential,
   TokenCredential,
   WebSocketOptions
 } from "@azure/core-amqp";
 import { ConnectionContext } from "./connectionContext";
 import { UserAgentOptions } from "@azure/core-http";
+import { SharedKeyCredential } from "./servicebusSharedKeyCredential";
 
 /**
  * Describes the options that can be provided while creating the ServiceBusClient.
