+++ sdk/servicebus/service-bus/src/util/sasServiceClientCredentials.ts
@@ -4,7 +4,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { AccessToken, SharedKeyCredential } from "@azure/core-amqp";
+import { AccessToken } from "@azure/core-amqp";
+import { SharedKeyCredential } from "../servicebusSharedKeyCredential";
 import { HttpHeaders, ServiceClientCredentials, WebResource } from "@azure/core-http";
 import { generateKey } from "./crypto";
